+++ system/dev/display/display/id-map.h
@@ -11,13 +11,13 @@ namespace display {
 // Helper for allowing structs which are identified by unique ids to be put in a hashmap.
 template <typename T> class IdMappable : public fbl::SinglyLinkedListable<T> {
 public:
-    int32_t id;
+    uint64_t id;
 
-    static size_t GetHash(int32_t id) { return id; }
+    static size_t GetHash(uint64_t id) { return id; }
 
-    int32_t GetKey() const { return id; }
+    uint64_t GetKey() const { return id; }
 
-    using Map = fbl::HashTable<int32_t, T>;
+    using Map = fbl::HashTable<uint64_t, T>;
 };
 
 } // namespace display
